+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/account/history/HistoryTestSetup.kt
@@ -29,11 +29,12 @@ class HistoryTestSetup(
         composeTestRule.setContent {
             ZcashTheme {
                 HistoryContainer(
-                    transactionState = initialHistoryUiState,
+                    isHideBalances = false,
                     onStatusClick = {},
                     onTransactionItemAction = {
                         onItemIdClickCount.incrementAndGet()
                     },
+                    transactionState = initialHistoryUiState,
                     walletRestoringState = WalletRestoringState.NONE,
                     walletSnapshot = WalletSnapshotFixture.new()
                 )