+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/balances/BalancesTestSetup.kt
@@ -38,6 +38,8 @@ class BalancesTestSetup(
                 onSettingsCount.incrementAndGet()
             },
             hideStatusDialog = {},
+            isHideBalances = false,
+            onHideBalances = {},
             showStatusDialog = null,
             onStatusClick = {},
             snackbarHostState = SnackbarHostState(),