+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/send/SendViewTestSetup.kt
@@ -114,6 +114,8 @@ class SendViewTestSetup(
                     // TODO [#1194]: Cover Current balances UI widget with tests
                     // TODO [#1194]: https://github.com/Electric-Coin-Company/zashi-android/issues/1194
                 },
+                isHideBalances = false,
+                onHideBalances = {},
                 hasCameraFeature = hasCameraFeature,
                 recipientAddressState = RecipientAddressState("", AddressType.Invalid()),
                 onRecipientAddressChange = {