+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/common/test/CommonTag.kt
@@ -6,4 +6,5 @@ package co.electriccoin.zcash.ui.common.test
 object CommonTag {
     const val WALLET_BIRTHDAY = "wallet_birthday"
     const val SETTINGS_TOP_BAR_BUTTON = "settings_top_bar_button"
+    const val HIDE_BALANCES_TOP_BAR_BUTTON = "hide_balances_top_bar_button"
 }