+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/common/viewmodel/HomeViewModel.kt
@@ -40,6 +40,27 @@ class HomeViewModel(application: Application) : AndroidViewModel(application) {
         setBooleanPreference(StandardPreferenceKeys.IS_RESTORING_INITIAL_WARNING_SEEN, true)
     }
 
+    /**
+     * A flow of the wallet balances visibility.
+     */
+    val isHideBalances: StateFlow<Boolean?> = booleanStateFlow(StandardPreferenceKeys.IS_HIDE_BALANCES)
+
+    fun setHideBalances(value: Boolean) {
+        setBooleanPreference(StandardPreferenceKeys.IS_HIDE_BALANCES, value)
+    }
+
+    val configurationFlow: StateFlow<Configuration?> =
+        AndroidConfigurationFactory.getInstance(application).getConfigurationFlow()
+            .stateIn(
+                viewModelScope,
+                SharingStarted.WhileSubscribed(ANDROID_STATE_FLOW_TIMEOUT.inWholeMilliseconds),
+                null
+            )
+
+    //
+    // PRIVATE HELPERS
+    //
+
     private fun booleanStateFlow(default: BooleanPreferenceDefault): StateFlow<Boolean?> =
         flow<Boolean?> {
             val preferenceProvider = StandardPreferenceSingleton.getInstance(getApplication())
@@ -49,14 +70,6 @@ class HomeViewModel(application: Application) : AndroidViewModel(application) {
             SharingStarted.WhileSubscribed(ANDROID_STATE_FLOW_TIMEOUT),
             null
         )
-
-    val configurationFlow: StateFlow<Configuration?> =
-        AndroidConfigurationFactory.getInstance(application).getConfigurationFlow()
-            .stateIn(
-                viewModelScope,
-                SharingStarted.WhileSubscribed(ANDROID_STATE_FLOW_TIMEOUT.inWholeMilliseconds),
-                null
-            )
 
     private fun setBooleanPreference(
         default: BooleanPreferenceDefault,